+++ src/vue/assets/Logo.vue
@@ -1,6 +1,6 @@
 <template>
   <router-link class="logotype" :to="vueRoutes.app">
-    <img :src="logoUrl" alt="TokenD">
+    <img :src="logoUrl" alt="Conto">
   </router-link>
 </template>
 
@@ -8,7 +8,7 @@
 import { vueRoutes } from '@/vue-router/routes'
 import { SchemeRegistry } from '@/modules-arch/scheme-registry'
 
-const DEFAULT_LOGO_URL = '/static/logo.svg'
+const DEFAULT_LOGO_URL = '/static/conto-logo.png'
 
 export default {
   data: _ => ({
@@ -26,5 +26,8 @@ export default {
 <style lang="scss">
 .logotype img {
   max-height: 4.5rem;
+  height: inherit;
+  width: inherit;
+  margin-left: -1.1rem;
 }
 </style>